+++ NEWS.md
@@ -4,9 +4,9 @@
 
 ### Added
 
-* `optional_build` option to allow the build task to complete even if Cargo is unavailable
-  and `github_releases` are disabled or unavailable. Useful for projects where fallback
-  code exists or where native extension code is desirable but not required.
+* `optional_rust_extension` option - prints a warning to STDERR instead of raising an exception, if
+  Cargo is unavailable and `github_releases` is either disabled or unavailable. Useful for projects
+  where either fallback code exists, or a native extension is desirable but not required. (#4, #6)
 
 ## [0.5.0] - 2016-07-18